- Unificación de funciones a PluginManager
